==> 1.Entrenamiento modelo sin usar reglas de lipinski/06_script-graficador-modelo-sin-usar-reglas-de-lipinski .ipynb <==
# Graficador del modelo sin usar reglas de lipinski
Una vez validado el modelo, se procede a graficar los datos del conjunto Activo"1" e Inactivo"0"


## Requisitos previos:
Asegúrate de tener un entorno de Python con las siguientes bibliotecas instaladas: matplotlib, pandas, seaborn, scikit-learn, imbalanced-learn, optuna.

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import matplotlib.cm as cm
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score
from sklearn.metrics import recall_score
from sklearn.model_selection import train_test_split
from sklearn.model_selection import train_test_split,cross_val_score,StratifiedKFold
from sklearn.metrics import confusion_matrix,classification_report,precision_score, recall_score, f1_score, accuracy_score
from sklearn.ensemble import RandomForestClassifier
from bayes_opt import BayesianOptimization
from sklearn import metrics
import pickle
import optuna
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import make_scorer, precision_score, recall_score, f1_score, confusion_matrix
from sklearn.model_selection import cross_validate
from sklearn.metrics import precision_score, recall_score, f1_score
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.metrics import confusion_matrix, balanced_accuracy_score, matthews_corrcoef
from sklearn.preprocessing import LabelEncoder
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import precision_score, recall_score, f1_score, balanced_accuracy_score, matthews_corrcoef
from sklearn.model_selection import cross_validate, train_test_split
import os
import pickle
import numpy as np
from sklearn.preprocessing import LabelEncoder
import matplotlib.pyplot as plt

## Generación graficas conjunto activo "1"

Este código realiza varias operaciones sobre un modelo entrenado y datos de validación y entrenamiento, así como visualizaciones de los resultados en forma de histogramas.

Primero, cambia el directorio de trabajo y carga datos de validación y objetivo desde archivos pickle. Utiliza un LabelEncoder para codificar los datos de validación y objetivo. Luego, carga un modelo entrenado desde un archivo pickle y utiliza este modelo para hacer predicciones sobre los datos de validación, obteniendo probabilidades de ser inactivo y predicciones de clasificación. Luego, filtra las muestras inactivas correctamente clasificadas e incorrectamente clasificadas.

Después, crea un histograma que muestra la distribución de las probabilidades de ser inactivo para las muestras inactivas correctamente clasificadas e incorrectamente clasificadas en los datos de validación. Guarda esta visualización en un archivo de imagen.

Luego, realiza operaciones similares para los datos de entrenamiento. Carga los datos de entrenamiento desde un archivo pickle, codifica los datos de entrenamiento y carga el modelo entrenado. Utiliza este modelo para hacer predicciones sobre los datos de entrenamiento, obteniendo probabilidades de ser inactivo y predicciones de clasificación. Filtra las muestras inactivas correctamente clasificadas e incorrectamente clasificadas y crea otro histograma similar al anterior. Finalmente, guarda esta visualización en otro archivo de imagen.

os.chdir('/home/aamaya/BACE1')

with open('validacionSIN.pickle', 'rb') as handle:
    validation_data = pickle.load(handle)


with open('df_targetSIN.pickle', 'rb') as handle:
    target_data = pickle.load(handle)

le = LabelEncoder()


X_val = validation_data.np_bv.apply(lambda x: le.fit_transform(list(x))).tolist()
y_val = validation_data.active


X_train = target_data.np_bv.apply(lambda x: le.transform(list(x))).tolist()
y_train = target_data.active


with open('modelo_entrenadoSIN.pickle', 'rb') as handle:
    model = pickle.load(handle)

y_prob = model.predict_proba(X_val)

y_prob_inactive = y_prob[y_val == 1, 1]  
y_pred = model.predict(X_val)  


correct_inactive = y_prob_inactive[y_pred[y_val == 1] == 1]

incorrect_inactive = y_prob_inactive[y_pred[y_val == 1] == 0]


fig, ax = plt.subplots()


range_correct = (0, 1)  # Full range from 0 to 1


n_correct, bins_correct, patches_correct = ax.hist(correct_inactive, bins=20, range=range_correct, alpha=0.5, color='green', label='active')

n_incorrect, bins_incorrect, patches_incorrect = ax.hist(incorrect_inactive, bins=20, range=range_correct, alpha=0.5, color='green')

ax.set_xlabel('Probabilidad de ser activos')
ax.set_ylabel('Numero de compuestos')


plt.title(f"A) Histograma de compuestos activos sin aplicar las reglas de Lipinski\n"
          f"(n = {np.sum(y_val == 1)})")


xticks = np.arange(0, 1.2, 0.2)  
ax.set_xticks(xticks)


xticklabels = [f'{x:.2f}' for x in xticks]
ax.set_xticklabels(xticklabels)


y_max = max(np.max(n_correct), np.max(n_incorrect))
ax.set_ylim(0, y_max + 40)  # Adding a buffer of 5 to the y-axis upper limit


plt.savefig("histogram_validation_1SIN.png")


import os
import pickle
import numpy as np
from sklearn.preprocessing import LabelEncoder
import matplotlib.pyplot as plt




with open('df_targetSIN.pickle', 'rb') as handle:
    target_data = pickle.load(handle)

le = LabelEncoder()


X_train = target_data.np_bv.apply(lambda x: le.fit_transform(list(x))).tolist()
y_train = target_data.active


with open('modelo_entrenadoSIN.pickle', 'rb') as handle:
    model = pickle.load(handle)

y_prob_train = model.predict_proba(X_train)

y_prob_inactive_train = y_prob_train[y_train == 1, 1]  # Probability of being inactive for inactive samples
y_pred_train = model.predict(X_train)  # Predicted classification (0 for inactive, 1 for active)


correct_inactive_train = y_prob_inactive_train[y_pred_train[y_train == 1] == 1]

incorrect_inactive_train = y_prob_inactive_train[y_pred_train[y_train == 1] == 0]


fig, ax = plt.subplots()


range_correct = (0, 1)  


n_correct, bins_correct, patches_correct = ax.hist(correct_inactive_train, bins=20, range=range_correct, alpha=0.5, color='green', label='active')

n_incorrect, bins_incorrect, patches_incorrect = ax.hist(incorrect_inactive_train, bins=20, range=range_correct, alpha=0.5, color='green')

ax.set_xlabel('Probabilidad de ser activos')
ax.set_ylabel('Numero de compuestos')


plt.title(f"A) Histograma de compuestos activos aplicando las reglas de Lipinski\n"
          f"(n = {np.sum(y_train == 1)})")


xticks = np.arange(0, 1.2, 0.2)  
ax.set_xticks(xticks)


xticklabels = [f'{x:.2f}' for x in xticks]
ax.set_xticklabels(xticklabels)


y_max = max(np.max(n_correct), np.max(n_incorrect))
ax.set_ylim(0, y_max + 100)  


plt.savefig("histogram_validation_1trainSIN.png")

## Generación graficas conjunto inactivo "0"

Este código realiza varias operaciones sobre un modelo entrenado y datos de validación y entrenamiento, así como visualizaciones de los resultados en forma de histogramas.

Primero, cambia el directorio de trabajo y carga datos de validación y objetivo desde archivos pickle. Utiliza un LabelEncoder para codificar los datos de validación y objetivo. Luego, carga un modelo entrenado desde un archivo pickle y utiliza este modelo para hacer predicciones sobre los datos de validación, obteniendo probabilidades de ser inactivo y predicciones de clasificación. Luego, filtra las muestras inactivas correctamente clasificadas e incorrectamente clasificadas.

Después, crea un histograma que muestra la distribución de las probabilidades de ser inactivo para las muestras inactivas correctamente clasificadas e incorrectamente clasificadas en los datos de validación. Guarda esta visualización en un archivo de imagen.

Luego, realiza operaciones similares para los datos de entrenamiento. Carga los datos de entrenamiento desde un archivo pickle, codifica los datos de entrenamiento y carga el modelo entrenado. Utiliza este modelo para hacer predicciones sobre los datos de entrenamiento, obteniendo probabilidades de ser inactivo y predicciones de clasificación. Filtra las muestras inactivas correctamente clasificadas e incorrectamente clasificadas y crea otro histograma similar al anterior. Finalmente, guarda esta visualización en otro archivo de imagen.

os.chdir('/home/aamaya/BACE1')


with open('validacionSIN.pickle', 'rb') as handle:
    validation_data = pickle.load(handle)


with open('df_targetSIN.pickle', 'rb') as handle:
    target_data = pickle.load(handle)

le = LabelEncoder()


X_val = validation_data.np_bv.apply(lambda x: le.fit_transform(list(x))).tolist()
y_val = validation_data.active


X_train = target_data.np_bv.apply(lambda x: le.transform(list(x))).tolist()
y_train = target_data.active


with open('modelo_entrenadoSIN.pickle', 'rb') as handle:
    model = pickle.load(handle)

y_prob = model.predict_proba(X_val)

y_prob_inactive = y_prob[y_val == 0, 0]  
y_pred = model.predict(X_val)  


correct_inactive = y_prob_inactive[y_pred[y_val == 0] == 0]

incorrect_inactive = y_prob_inactive[y_pred[y_val == 0] == 1]


fig, ax = plt.subplots()


range_correct = (0, 1) 

n_correct, bins_correct, patches_correct = ax.hist(1 - correct_inactive, bins=20, range=range_correct, alpha=0.5, color='red', label='Inactive')

n_incorrect, bins_incorrect, patches_incorrect = ax.hist(1 - incorrect_inactive, bins=20, range=range_correct, alpha=0.5, color='red')

ax.set_xlabel('Probabilidad de ser activos')
ax.set_ylabel('Numero de compuestos')


plt.title(f"B) Histograma de compuestos inactivos sin aplicar las reglas de Lipinski\n"
          f"(n = {np.sum(y_val == 0)})")


xticks = np.arange(0, 1.2, 0.2)  
ax.set_xticks(xticks)


xticklabels = [f'{x:.2f}' for x in xticks]
ax.set_xticklabels(xticklabels)


y_max = max(np.max(n_correct), np.max(n_incorrect))
ax.set_ylim(0, y_max + 40)  


plt.savefig("histogram_validation_0SIN.png")

import os
import pickle
import numpy as np
from sklearn.preprocessing import LabelEncoder
import matplotlib.pyplot as plt




with open('df_targetSIN.pickle', 'rb') as handle:
    target_data = pickle.load(handle)

le = LabelEncoder()


X_train = target_data.np_bv.apply(lambda x: le.fit_transform(list(x))).tolist()
y_train = target_data.active


with open('modelo_entrenadoSIN.pickle', 'rb') as handle:
    model = pickle.load(handle)

y_prob_train = model.predict_proba(X_train)

y_prob_inactive_train = y_prob_train[y_train == 0, 0]  
y_pred_train = model.predict(X_train)  


correct_inactive_train = y_prob_inactive_train[y_pred_train[y_train == 0] == 0]

incorrect_inactive_train = y_prob_inactive_train[y_pred_train[y_train == 0] == 1]


fig, ax = plt.subplots()


range_correct = (0, 1) 


n_correct, bins_correct, patches_correct = ax.hist(1-correct_inactive_train, bins=20, range=range_correct, alpha=0.5, color='red', label='active')

n_incorrect, bins_incorrect, patches_incorrect = ax.hist(1-incorrect_inactive_train, bins=20, range=range_correct, alpha=0.5, color='red')

ax.set_xlabel('Probabilidad de ser activos')
ax.set_ylabel('Numero de compuestos')


plt.title(f"B) Histograma de compuestos inactivos sin aplicar las reglas de Lipinski\n"
          f"(n = {np.sum(y_train == 0)})")


xticks = np.arange(0, 1.2, 0.2)  
ax.set_xticks(xticks)


xticklabels = [f'{x:.2f}' for x in xticks]
ax.set_xticklabels(xticklabels)


y_max = max(np.max(n_correct), np.max(n_incorrect))
ax.set_ylim(0, y_max + 100)  


plt.savefig("histogram_validation_0trainSIN.png")


Script elaborado por Nelson Alejandro Amaya Orozco, como trabajo de grado, para el grupo de investigacion RamirezLAB
